+++ django_webdav_storage/apps.py
@@ -1,0 +1,5 @@
+from django import apps
+
+
+class DjangoWebDavStorageConfig(apps.AppConfig):
+    name = 'django_webdav_storage'