+++ example_project/example_app/admin.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from .models import File
+
+
+@admin.register(File)
+class FileAdmin(admin.ModelAdmin):
+    pass