+++ example_project/example_app/forms.py
@@ -1,0 +1,9 @@
+from django import forms
+
+from .models import File
+
+
+class FileForm(forms.ModelForm):
+    class Meta:
+        model = File
+        fields = '__all__'