+++ example_project/example_app/migrations/0001_initial.py
@@ -1,0 +1,38 @@
+from django.db import migrations, models
+
+import example_app.models
+
+
+class Migration(migrations.Migration):
+
+    initial = True
+
+    dependencies = [
+    ]
+
+    operations = [
+        migrations.CreateModel(
+            name='File',
+            fields=[
+                (
+                    'id',
+                    models.AutoField(
+                        auto_created=True,
+                        primary_key=True,
+                        serialize=False,
+                        verbose_name='ID'
+                    )
+                ),
+                (
+                    'file',
+                    models.FileField(
+                        upload_to=example_app.models.upload_to
+                    )
+                ),
+            ],
+            options={
+                'verbose_name': 'file',
+                'verbose_name_plural': 'files',
+            },
+        ),
+    ]